Move methods and extensions on EventRegistration and EventConsumerRegistration to DefaultEventConfigurator
